requirements.txt
names
dj-database-url==0.4.1
Django==1.9.7
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==2.0.6
newrelic
confluent-kafka
kafka-helper
djangorestframework
markdown
django-filter

cffi==1.8.2
cryptography
enum34==1.1.6
idna==2.1
ipaddress==1.0.17
pyasn1==0.1.9
pycparser==2.14

django-cors-headers
